Select correct disk for HANA when BACKEND is not qemu
For the HANA cli installation in qemu, we expect the job to have
correctly configured settings so a second disk with enough space
for HANA is configured in the VM. When backend is not qemu, the
test will rely on whatever disks are already configured in the
system, be it bare metal or a ppc64le LPAR; in these cases, the
tests selects /dev/sdb by default, even if this device does not
have enough space for HANA or if it does not exist.

This commit introduces changes to make the test module probe for
available block devices, and verify if these have enough space for
HANA as calculated from the system RAM. If no disks are found with
enough space for HANA, test will fail.

Since SLES4SAP is installed by default with LVM for SWAP and root FS,
the test module will filter out block devices that are configured as
LVM physical volumes. It will also check for multipath devices if the
installation was done with multipath configured. This is determined by
the MULTIPATH=1 and MULTIPATH_CONFIRM=yes settings.

Test module depends on the output of the lsblk command to probe devices
and get their size.